extract mongooseoptions
